+++ packages/lumx-vue/src/components/grid-column/index.ts
@@ -1,0 +1,3 @@
+export { default as GridColumn } from './GridColumn';
+export type { GridColumnProps, GridColumnGapSize } from './GridColumn';
+export { CLASSNAME as GRID_COLUMN_CLASSNAME, COMPONENT_NAME as GRID_COLUMN_COMPONENT_NAME } from './GridColumn';