+++ packages/site-demo/src/components/content/Placeholder/index.tsx
@@ -1,12 +1,12 @@
 import type React from 'react';
 
-type Props = React.FC<{
+type Props = {
     name: string;
     height?: React.CSSProperties['height'];
     width?: React.CSSProperties['width'];
     textAlign?: 'center' | 'right' | 'left';
-}>;
-export const Placeholder: Props = ({ name, height = 36, width = 36, textAlign = 'center' }) => (
+};
+export const Placeholder: React.FC<Props> = ({ name, height = 36, width = 36, textAlign = 'center' }) => (
     <div
         style={{
             height,