+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
     -   Create the `Link` component
     -   Create the `Divider` component
     -   Create the `Thumbnail` component
+    -   Create the `GridColumn` component
 
 ### Changed
 
@@ -26,6 +27,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
     -   Moved `Link` from `@lumx/react`
     -   Moved `Divider` from `@lumx/react`
     -   Moved `Thumbnail` from `@lumx/react`
+    -   Moved `GridColumn` from `@lumx/react`
     -   Uniformize documentation of `disabled` (as an alias to `isDisabled`) props across core components
 
 ## [4.3.1][] - 2026-02-11